finishi clue convert transaction

+++ src/main/java/com/awei/crm/service/impl/ClueServiceImpl.java
@@ -1,5 +1,6 @@
 package com.awei.crm.service.impl;
 
+import com.awei.crm.exception.CULDException;
 import com.awei.crm.mapper.*;
 import com.awei.crm.model.*;
 import com.awei.crm.service.ClueService;
@@ -7,6 +8,9 @@ import com.awei.crm.utils.DateTimeUtil;
 import com.awei.crm.utils.UUIDUtil;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Isolation;
+import org.springframework.transaction.annotation.Propagation;
+import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
 import java.util.Map;
@@ -76,8 +80,15 @@ public class ClueServiceImpl implements ClueService {
         return clueMapper.deleteByPrimaryKey(clueId);
     }
 
+    @Transactional(
+
+            propagation = Propagation.REQUIRED,
+            rollbackFor = {NullPointerException.class,CULDException.class},
+            isolation = Isolation.DEFAULT
+    )
     @Override
-    public boolean convert(Tran tran, User user, String clueId) {
+    public boolean convert(Tran tran, User user, String clueId) throws CULDException {
+
         //(1) 获取到线索id，通过线索id获取线索对象（线索对象当中封装了线索的信息
         Clue c = clueMapper.selectByPrimaryKey(clueId);
         //(2) 通过线索对象提取客户信息，当该客户不存在的时候，新建客户（根据公司的名称精确匹配，判断该客户是否存在！）
@@ -96,7 +107,11 @@ public class ClueServiceImpl implements ClueService {
             customer.setContactSummary(c.getContactSummary());
             customer.setAddress(c.getAddress());
             //新建 客户
-            customerMapper.insert(customer);
+           int insertNum = customerMapper.insert(customer);
+            if (insertNum ==0) {
+                //插入失败
+                throw new CULDException("添加客户信息失败,客户ID：" + customer.getId());
+            }
         }
         //(3) 通过线索对象提取联系人信息，保存联系人
         Contacts contacts = new Contacts();
@@ -114,9 +129,12 @@ public class ClueServiceImpl implements ClueService {
         contacts.setMphone(c.getMphone());
         contacts.setJob(c.getJob());
         contacts.setCustomerId(customer.getId());
-        contactsMapper.insert(contacts);
+        int insertNum = contactsMapper.insert(contacts);
+        if (insertNum == 0) {
+            throw new CULDException("添加联系人信息失败，联系人ID：" + contacts.getId());
+        }
         //(4) 线索备注转换到客户备注以及联系人备注
-        List<ClueRemark> clRmkList= clueRemarkMapper.selectByClueId(clueId);
+        List<ClueRemark> clRmkList = clueRemarkMapper.selectByClueId(clueId);
         if (null != clRmkList) {
             for (ClueRemark clueRemark : clRmkList) {
                 CustomerRemark customerRemark = new CustomerRemark();
@@ -137,12 +155,15 @@ public class ClueServiceImpl implements ClueService {
 
                 contactsRemark.setEditFlag("0");
                 customerRemark.setEditFlag("0");
-                customerRemarkMapper.insert(customerRemark);
-                contactsRemarkMapper.insert(contactsRemark);
+                insertNum= customerRemarkMapper.insert(customerRemark);
+                if (insertNum == 0) {
+                    throw new CULDException("添加客户备注失败,客户ID：" + customerRemark.getCustomerId());
+                }
+                insertNum=contactsRemarkMapper.insert(contactsRemark);
+                if (insertNum == 0) {
+                    throw new CULDException("添加联系人备注失败联系人ID：" + contactsRemark.getContactsId());
+                }
             }
-            //干掉与该条线索相关的备注
-            clueRemarkMapper.deleteByClueId(clueId);
-
         }
 
 
@@ -154,10 +175,12 @@ public class ClueServiceImpl implements ClueService {
                 contactsActivityRelation.setActivityId(re.getActivityId());
                 contactsActivityRelation.setContactsId(re.getClueId());
                 contactsActivityRelation.setId(UUIDUtil.getUUID());
-                contactsActivityRelationMapper.insert(contactsActivityRelation);
+                insertNum = contactsActivityRelationMapper.insert(contactsActivityRelation);
+                if (insertNum == 0) {
+                    throw new CULDException("添加联系人与市场活动的关系失败,关系ID： " + contactsActivityRelation.getId());
+                }
+
             }
-            //把线索和市场活动 的关系 干掉
-            clueActivityRelationMapper.deleteByClueId(clueId);
         }
 
 
@@ -169,7 +192,10 @@ public class ClueServiceImpl implements ClueService {
             tran.setCreateBy(user.getId());
             tran.setContactsId(contacts.getId());
             tran.setCustomerId(customer.getId());
-            tranMapper.insert(tran);
+            insertNum = tranMapper.insert(tran);
+            if (insertNum == 0) {
+                throw new CULDException("添加交易失败,交易客户ID：" + tran.getCustomerId());
+            }
             //(7) 如果创建了交易，则创建一条该交易下的交易历史
             TranHistory history = new TranHistory();
             history.setId(UUIDUtil.getUUID());
@@ -179,16 +205,33 @@ public class ClueServiceImpl implements ClueService {
             history.setMoney(tran.getMoney());
             history.setStage(tran.getStage());
             history.setExpectedDate(tran.getExpectedDate());
-            tranHistoryMapper.insert(history);
+            insertNum = tranHistoryMapper.insert(history);
+            if (insertNum == 0) {
+                throw new CULDException("添加交易历史记录失败,交易ID：" + tran.getId());
+            }
         }
         //(8) 删除线索备注
-        int clueDeleteNum = clueRemarkMapper.deleteByClueId(clueId);
+        List<ClueRemark> crList = clueRemarkMapper.selectByClueId(clueId);
+        if (crList.size()!=0) {
+            //不为空，说明该线索有备注，把他干掉
+            int deleteNum = clueRemarkMapper.deleteByClueId(clueId);
+            if (deleteNum == 0) {
+                throw new CULDException("删除该线索的备注失败,线索ID："+clueId);
+            }
+        }
         //(9) 删除线索和市场活动的关系
-        int clueActRelationNum = clueActivityRelationMapper.deleteByClueId(clueId);
+        List<ClueActivityRelation> carList = clueActivityRelationMapper.selectByClueId(clueId);
+        if (carList.size()!=0) {
+            int deleteNum = clueActivityRelationMapper.deleteByClueId(clueId);
+            if (deleteNum == 0) {
+                throw new CULDException("删除线索跟市场活动的关系失败，线索ID：" + clueId);
+            }
+        }
         //(10) 删除线索
         int deleteClueNum = clueMapper.deleteByPrimaryKey(clueId);
-
-
+        if (deleteClueNum == 0) {
+            throw new CULDException("删除线索失败，线索ID:" + clueId);
+        }
         return true;
     }
